+++ components/atom/helpText/src/index.js
@@ -6,11 +6,11 @@ import Injector from '@s-ui/react-primitive-injector'
 
 import {BASE_CLASS} from './settings.js'
 
-const AtomHelpText = forwardRef(({text}, forwardedRef) => {
+const AtomHelpText = forwardRef(({text, ...props}, forwardedRef) => {
   const isTextString = typeof text === 'string'
   const Component = isTextString ? 'span' : Injector
   return (
-    <Component className={BASE_CLASS} {...(isTextString && {ref: forwardedRef})}>
+    <Component className={BASE_CLASS} {...props} {...(isTextString && {ref: forwardedRef})}>
       {text}
     </Component>
   )